+++ tests/Fixtures/EnumThatExtendsValidObjectsEnum.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Zlikavac32\Enum\Tests\Fixtures;
+
+/**
+ * @method static EnumThatExtendsValidObjectsEnum ENUM_A
+ */
+abstract class EnumThatExtendsValidObjectsEnum extends ValidObjectsEnum
+{
+    protected static function enumerate(): array
+    {
+        return [
+            'ENUM_A' => new class extends EnumThatExtendsValidObjectsEnum {}
+        ];
+    }
+}