+++ tests/Fixtures/EnumThatExtendsValidStringEnum.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Zlikavac32\Enum\Tests\Fixtures;
+
+/**
+ * @method static EnumThatExtendsValidStringEnum ENUM_A
+ */
+abstract class EnumThatExtendsValidStringEnum extends ValidStringEnum
+{
+
+}