+++ tests/Fixtures/NonObjectEnumerationObjectsEnum.php
@@ -6,6 +6,9 @@ namespace Zlikavac32\Enum\Tests\Fixtures;
 
 use Zlikavac32\Enum\Enum;
 
+/**
+ * @method static NonObjectEnumerationObjectsEnum ALIAS
+ */
 abstract class NonObjectEnumerationObjectsEnum extends Enum
 {
     protected static function enumerate(): array